二次修改

+++ src/com/ntu/oa/dao/DynamicDao.java
@@ -8,5 +8,5 @@ public interface DynamicDao {
 	
 	void upDynanmic(Map<String, Object> info);
 	
-	void delDynamic(int id);
+	void delDynamic(long id);
 }
